requirements.txt
camelot-py[cv]
pdfplumber
reportlab
langchain_groq
load_dotenv
langchain_core
streamlit
crewai
langchain
ipdb
requests
crewai-tools
openai
matplotlib
pyrebase4
firebase-admin
openpyxl
plotly
python-multipart